improved random pixel selection to paint more efficiently
instead of randomly picking a pair of x,y coordinates to paint, generate
a list of random indices as seeds and loop through each of them. this
results in the picture being painted more quickly.
also, changed requested image to be half size as it is not necessary to
use native resolution